General improvements to MapTool and connection startup
When starting a personal server or connecting to a remote server, verify that we don't have a server currently
running. This used to only be done when starting a hosted server. The existing and new checks report failure through a
logger rather than dumping the stack to stderr.

Personal server startup is more like a real server by using `installClient()`. They also have their icon once again set
to indicate hosting rather than disconnected. We also wait on personal server starts via futures to avoid sequencing
issues during app startup.

Handshake completion callbacks now provide a flag indicating whether the handshake was successful. This is so we can
immediately bail on new connections rather than assuming they are successful.

Since `MapTool.startServer()` always copies the campaign, we have no need for the parameter to control that. Moreover,
we stash the current campaign in the client so that we can better recover if the local handshake fails for whatever
reason.

The default topology types from preferences are now applied directly to the initial basic campaign. This avoids the need
to look up any zone renderers, which is a dubious task so early in the init process.

Personal server be waiting on via future, no blocking