Big ol' refactor to MapTool connections
`MapToolServerConnection` has been merged into `MapToolServer`. The responsibilities were not clear and everything gets
a whole lot easier if they are the same thing. If they are to be separated out again in the future,
`MapToolServerConnection` should deal in connections, while `MapToolServer` should deal in players.

Similarly, `MapToolConnection` was merged into `MapToolClient`. Essentially everything that `MapToolConnection` was
doing is either of direct concern to `MapToolClient` or can be done directly with the underlying `Connection`.

Personal servers have been reunited with hosted servers, both being represented by `MapToolServer`
again. `MapToolServer` now has the concept of a "local client" which is trivially connected to it through a
`DirectConnection`. This allows both personal servers and local hosted servers to start instantly and with low overhead
for the local client, but without otherwise treating local clients specially. Along with this, no handshake is expected
for local connections, because the server already knows the client that is connecting locally. Only remote clients have
to perform a handshake.