[async] Rework ImageManager to avoid giant critical sections
No more imageLoaderMutex since that locks the entire thing. Instead we have `CacheEntry` which brings together the
loaded image (if any) and the list of observers that still need to be notified. This one type takes things under a lock
only as needed, and is restricted to doing so for a specific entry, no the entire manager. It also allows notifications
to execute outside the critical section, meaning there should be no chance of a deadlock.

The new critical sections are extremely small and boil down to a collection read or write, and possibly a task
submission. I still feel like they could be a bit smaller, but it's still an improvement.

On the other hand, `BackgroundImageLoader` no longer checks to see if another loader has already completed in its run()
method. This means that if two happen to be submitted for the same asset, both will run to completion. It may seem like
a small thing to just check, but doing it the new way avoids concurrency concerns with reading from the entry. And it
should also never happen that the same asset is loaded again (unless it was flushed) so this is not much of a concern.