Build out encapsulated asset caches
They mainly deal in `LazyAsset`, a new type that represents an asset that has yet to be loaded. It is essentially a
`AssetHeader` pair with a strategy for loading the asset.

Also included is a "layered cache" which cna layer any number of caches on top of each other. It will find an asset in
the fastest (first) cache available, and will populate faster caches when loading from slower caches.

Another utility is the "verifying loader". This double-checks the MD5 of an asset loaded from a cache and remove the
asset from the cache if it doesn't match.