Make initialize list rendering more consistent
The token image size is no longer dependent on whether states are being rendered. It is also not dependent on the number
of lines of text, but will always be drawn at 50x50 pixels.

The name plate was adjusted, too. It now accounts for insets (previously only done for label preferred size), and the
insets have been expanded slightly as the text was too close to the border.