[expose] Strategize expose tools
Required a new `FreehandStrategy` and explict support for it in the new `ExposeTool`. Should propagate this to
`TopologyTool` despite not being an option right now. Like the polygon topology tool, the freehand expose tool now
renders its fill so the exposed area is obvious to the user.

Also required updating `OvalStrategy` to allow for smooth ovals. The stepped variant is useful to keep topology points
down (though that may no longer be important).

TODO Expose tools are supposed to show measurements. Not sure why. On the other hand, we could say it is weird that
topology tools don't show measurements. Either way, we'll need to devise a scheme for them. The rules are:
- Rectangle, oval tool show measures of the rectangular bounds.
- Iso rectangle shows measures of the iso shape side length.
- Polygon shows length of current segment.
- Freehand does not show measure.