Replace existing topology tool implementations with one class
The new `TopologyTool` uses the previously introduced strategies to define all topology tools. They are instance-based,
using the recent supporting change to the toolbox.

`TopologyTool` merely interprets the results of the strategies. The strategies are responsible for defining the shapes,
but have no knowledge of what the shape will be used for.