Deprecate now-unused drawing types
`Oval`, `Rectangle`, and `Cross` no longer serve a purpose and so have been deprecated. It is possible for existing
campaigns to have these serialized in them, so we are keeping the structure around. However, each one will
`readResolve()` itself into a modern type.

Also be clear about the set of supported types in `ShapeDrawable`. The new `ShapeDrawable#getShapeTypeName()` outputs
one of `"Rectangle"`, `"Oval"`, `"Polygon"`, `"Area"`, or `"Unknown"`, which callers can use to make meaningful
decisions about the general type of shape wrapped in the drawable. This matters especially for ellipses, where several
places refers to the simple name of `Ellipse2D.Float`, which is just `"Float"`. Now they refer to `"Oval"`, or use
proper type checks where needed.