Refactor TokenVBL.getTopology_underToken()
The results of a few calculations were not used, so a good chunk of the method could just be deleted. Beyond that, the
method was written to be as condensed as possible, and avoiding unnecessary `Area` copies.

The method `TokenVBL.getMapTopology_transformed()` was removed because it had to do everything that
`TokenVBL.getTopology_underToken()` plus transforming it back into token space. A new method
`TokenVBL.transformTopology_toToken()` takes its place, and can transform an `Area` (no matter the source) into token
space. This process was taken from `TokenVBL.getMapTopology_transformed()`, but cleaned up to produce a single
transformation and avoid any intermediate `Area`.