Introduce new topology types
`MaskTopology` represents legacy masks as polygons. This is used in lighting and pathfinding algorithms, though the
original `Area` is maintained and serialized in the model for the sake of existing tools and macros.

`WallTopology` is a new form of topology, made of thin walls that form a graph. Edge wall (edge) blocks both vision and
movement from any direction. The plan is to extend this in the future to control vision and movement separately, and to
add directionality to the wall, but for now only the basic option is available.

`WallTopologyConverter` ensures that `WallTopology` can be serialized without exposing the implementation details, and
to make the XML look reasonable.