[type ]Introduce new topology types
`MaskTopology` represents for legacy masks as polygons. This is used in lighting and pathfinding algorithms, though the
original `Area` is maintained and serialized in the model.

`WallTopology` is a graph of thin walls that block vision and movement. Future improvements will refine these to support
more cases. It comes with a converter (`WallTopologyConverter`) to ensure that the XML remains reasonable and does not
depend on its implementation details.