[pathfinding] Use the new topologies in pathfinding
This uses the new noding support in `Zone`, which means `ZoneView` no longer has to cache the combination of map and
token topology. This does result in a slight difference in behaviour, but it is a proper change: token MBL is now
treated completely separately from map MBL and is not merged down into it.