Use the new topologies in A* pathfinding
This reads the applicable topologies directly from the `Zone`, which means the `ZoneView` no longer has to cache the
combination of map and token topologies. There is a slight change in behaviour here: before, the map + token topologies
would be merged together, then the key token's topology would be subtracted back out; now the map topology is only
merged with token topology other than the key token. The former behaviour would allow the key token to carve out a path
using it own topology, but now it can no longer do that. It also opens the door to a different caching option that can
be done entirely in the walker (this is not actually done yet).