Add tooling for wall topology
A new topology tool is dedicated to working with `WallTopology`. It will render legacy topology for reference, but can
only modify walls. Walls are drawn as stroked line segments with circular handles at the vertices. Both vertices and
entire walls can be dragged. Vertices can be merged with one another, and can split existing walls. Shift-clicking will
delete the hovered vertex or wall. If a vertex is deleted, all walls connected to it are deleted as well.

Operations are expressed through a new `Rig` type that understands `Handle` specifically and `Movable` as a more general
draggable concept. This rigging is only use for wall topology for now, but the hope is to one day apply it to drawings
and perhaps even legacy mask topology in order to make those conveniently editable.

Since topology type has no meaning for wall topology, the topology type selector is disabled when the wall topology tool
is selected.