Add data structures for prepared topologies
For vision, the `NodedTopology` offers a fully-noded copy of the input topologies, both legacy masks and walls. The
vision sweep requires endpoints to be place at every intersection, otherwise important steps are missed. While this was
implicitly doable with masks by treating each mask separately, we need to be explicit about it with the introduction of
walls since they can arbitrarily overlap with each other and with the masks.

For pathfinding, the `MovementBlockingTopology` abstracts the creation of `PreparedGeometry`. A* really needs good
`PreparedGeometry` to have acceptable performance, but this is only achievable if the input geometries are similar in
type. This was true when using legacy masks as the masks would convert into `Polygonal` geometry. But wall topology
converts to `Lineal` geometry, and mixing the two results in very poor performance. The `MovementBlockingTopology` will
create separate `PreparedGeometry` for the `Polygonal` and `Lineal` inputs. While this requires two separate
intersection checks, this is still much faster than trying to combine the two types.