Add support for landing maps
Each `Campaign` now stores the ID of its landing map, if one is set. The user sets one using from the app menu,
_Map > Set as landing map_. When a player client connects and the campaign has a landing map set, the client will look
up that landing map to display first.

This does not apply to GM clients. Also, the local client will prefer to stay on the current map if possible when
starting and stopping a server.