Add support in UVTT importer for unique lights
The importer will synthesize unique lights for each light in the UVTT file. The range will be taken from the light's
`"range"` field, and the lumens will be set to 100. The colour will be set to the light's `"color"` field, unless the
map image has baked lighting in - in that case, the unique light source will be clear. The `IGNORES-VBL` flag will be
set if the light's `"shadow"` flag is `false`.

All light properties are now written to the GM notes. Previously we wrote a hand-picked subset, but now everything is
there. This avoids any chance of missing some in the future if the UVTT format adds additional properties.

There are a few limitations with the import:
1. UVTT does not include the shape or texture of the light, so imported lights are always treated as circles.
2. We have no equivalent to "intensity" so that remains unused.