Tweak method in ServerCommand for attaching light sources
The new version is not an overload of `updateTokenProperty()` but has a more descriptive name. It also does not rely on
the caller deciding which `Token.Update` value to use. Instead, a `boolean` is accepted to toggle the light source on or
off, and the method will pick one of the two valid `Token.Update` value based on that.