Add support in existing macros for unique light sources
This only adds the ability to write a `"$token"` category in `getLights()`, `setLight()`, and `hasLightSource()` to
restrict the operation to unique light sources on the given token, rather than restricting it to a campaign category. It
also extends the wildcard (`"*"`) support in these functions to also look for unique light sources on the token.