Add support in the UI for unique light sources
The Edit Token dialog now sports a textbox where unique light sources can be defined. The syntax is identical to the light
syntax in the Campaign Properties dialog except that categories are not supported.

If a token has unique light sources, its context menu will display them under "Light Sources". They will be in their own
"Unique" category, separated from the campaign categories. The functionality is identical to that for campaign lights.