Make MapToolRegistry.getAddress return InetAddress
The parsed value is more directly useful for most callers
since we can assume it is a valid address.

Callers that want it as a string have to convert it to a string now
and handle that it might be null but that's easier than turning it into
the InetAddress from the String since InetAddress.getByName can raise an
exception, and CompletableFutire.thenApply can't take functions that
throw exceptions.