Move cached external addr getter out of UI
The MapToolServer class calling into the ConnectionInfoDialog class
to get a cached copy of the external address wasn't a clean abstraction.

This introduces a new NetUtil class because this already needs to be
done in two places and I intend to add a third.

Since a completed future holds the result of its computation
it is handy to keep a reference to it around rather than jump through
hoops to have it assign its result to the cache.