Implement missing reachability check
PR #4208 had a request to add a reachability check
on the grounds that if an address hasn't got a route to rptools.net
then it's probably not configured to be reachable from your peers
since your routing table is used to send packets back the other way.

This was not implemented correctly, it just added a check that
rptools.net resolved, which happens through the preferred interface
which is able to reach the configured DNS server
rather than the address being iterated over.

https://docs.oracle.com/en/java/javase/21/docs/api//java.base/java/net/InetAddress.html#isReachable(java.net.NetworkInterface,int,int)
could be used for this reachability check
but that's implemented as using IGMP (ping) where available
or falling back to a TCP connection to port 7.

Restrictive firewalling could cause this to fail but we know rptools.net
serves SSL, so we can connect to that and immediately close the
connection.