Simplify SelectionSet case work
The presence of the walker is used to indicated whether we are currently using pathfinding and snap-to-grid
behaviour. We no longer check that at every opportunity. Also, the task requires a non-null walker, which was not
checked.

This enables us to easily check against a null walker when it matters.