Add direction to walls
Each wall can have a main direction (`Left`, `Right`) or be bidirectional (`Both`). Relative to the main direction,
specific uses (`Sight`, `Light`, `Aura`) can be configured to respect the main direction (`SameDirection`), work in the
opposite direction (`ReverseDirection`), ignore directionality and be bidirectional (`ForceBoth`) or ignore the wall
entirely (`Disabled`). Movement also supports the `ForceBoth` and `Disabled` cases, but does not support direction
itself and so does not have the `SameDirection` and `ReverseDirection` cases.

Internally, the representation of walls got more complicated, but a future change will work to reduce that
again. `WallInternal` now sports a mutable `WallData` that is capable of being overwritten and merged with `WallData`
from other walls. Individual walls can have their data be updated through the new `UpdateWallMessage` message. The
`WallData` is what carries the directional information for the wall.

A new control panel gives access to changing wall properties. The currently selected wall will have its state reflected
in the control panel, and modifying the values there will update the wall. When drawing a new wall, the current values
from the control panel will be applied.

The XML for walls has been extended to store the wall's direction and modifiers as child elements.