Make MapToolServer's ID explicitly nullable
ServiceAnnouncer requires a non-null String, but an empty server ID is
not well-formed so a ServiceAnnouncer shouldn't be created.

This was previously handled by MapToolServer being passed null config
and the absence of a port to give was sufficient to prevent creating a
ServiceAnnouncer with an empty ID,
but we can be more defensive about this and check for a null pointer.