Announce bound instead of configured server port
If a port number of 0 is chosen then bind() will open a random port in
an ephemeral range instead of the provided port number
and this port can then be retrieved using getLocalPort().

This enables starting servers on arbitrary ports instead of
preconfigured ones.

Since only direct connections need to know the port number
connections using the registry or LAN can use an ephemeral port number
and the connection information tab can display the port direct
connections should use.

This requires a router with functioning UPnP instead of port forwarding
though so is not univerally useful.