Add LAN ID and connect link buttons to ConnectionInfoDialog
The button that looks like file copy copies the rptools-maptool URI to
the clipboard, and the icon is meant to suggest copying.

The button that looks like a cursor is meant to imply clicking,
which the copyable HTTP URI is because chat applications will
automatically turn that text into a link that can be clicked
but does require the web browser opening the page and forwarding to the
payload first.