Add property default and reset JS bindings
resetProperty was conspicuously absent
and is necessary to support resetting a token to default
which is convenient for things like a Full Heal or removing expired
temporary HP.

Another reason to reset properties is if they have been restored to
their default and you wish to reduce memory usage and serialization size
but to do so without coupling to the campaign requires functions to get
the default.

getPropertyDefault gets the default value from the properties verbatim
and getEvaluatedPropertDefault evaluates the expression in the context
of the token.