Set release asset names by copying them to another directory
This completely avoids potential name conflicts with renaming a file to the same name, or possibly just that another
asset might have that name already.

We also apply this to the uberJar to make sure its name always matches the expectation as well, and so that all assets
can be found in the same location.