Add a gradle task to zip the jpackage images for release
The new `zipJpackageImage` consumes the platform-specific images into archives that can be published as part of a
release. The archive name will include the version and the platform name ("win", "mac", or "linux"). Users can download
and unzip these archives to any location they wish to achieve parallel installations. This also offers an alternative to
the uber jar on all platforms for those not wishing to install via their system installers or package manager.

The publish workflow has been updated to invoke `zipJpackageImage` in addition to `jpackage` in order to make the files
available for releases.