Disable focusable tooltips on all RSyntaxTextArea
The default is for focusable tooltips to e be enabled. As it happens, only the _Tooltip_ field of the macro editor
dialog used tooltips, so it is the only one with an observable problem. But if we ever add tooltips to the others,
those would have the same issue. Hence this changes all uses of `RSyntaxTextArea`.

This applies to uses in `.form` fields and in `.java`.