Make the initial trivial campaign truly empty
This fixes #5219 by not loading `ZoneFactory` too early, as its static initializer can only succeed if the default
resources have already been installed. On MT's first launch, this is not true, so the asset is never made available to
new zones.

An added perk is that this trivial campaign is even more lightweight now, since there is no need to read, hash, and load
an image.