ZoneFactory to cache default grass texture only on success
This is a second complete fix for #5219. If the `ZoneFactory` fails to get the default texture, it will not cache that
result but will try again for the next zone that gets created. Additionally, if the texture was not loaded,
`ZoneFactory` will fall back to a solid color so that the user is not left with a screenfull of red X's.

This means that if the texture file is found but there is an error loading it, then `ZoneFactory` will try to read it
again the next time a zone is created. For a corrupt file, this could result in repeated failures and extra work
repeatedly trying to load the asset. However, zone creation is not a high-frequency event, so this is acceptable.