Recognise model/vnd.mts MIME as text
If an addon contains an empty script file it doesn't detect as text
and the .mts suffix heuristically makes it "model/vnd.mts".

This would get mapped to INVALID and thus a broken image
which would not have a string representation and when the add-on loader
attempts to load the asset as a string it would assert.