Do not use intermediate images for rendering tokens
It is very poor for performance and quality to produce a sequence of intermediate images for each transformation that
needs to be applied. With the right caches, the performance issue can be mitigated, but the quality will always be poor.

We now use the main token image, and apply a single `AffineTranformation` to it while rendering to the screen. This
preserves maximum quality, and is quite performant as well.