Clean up and simplify logic in TokenRenderer
There are no more fields devoted to the current token. These are instead passed a method parameters.
- `opacity`
- `canSpin`
- `isoFigure`
- `renderImage`
- `position`
- `token`

Also remove the `renderImageMap` in preparation for realtime transformation coming soon. With that gone, there is no
more need for `tokenStateMap`, `compareStates()` or `TokenState`, nor does `TokenRenderer` need to invalidate itself in
response to any events. NOTE: this is very bad for performance, and is only okay because we're about to revamp how token
rendering works.