Work with the Gdx-provided back buffer size for HiDPI support
The width and height provided to `GdxRenderer#resize(int, int)` are incorrect in every way. By looking up the back
buffer size diretly from `Gdx.graphics`, we are able to get sane rendering.

We also split the hud cam from the blit cam. The blit cam is for compositing layers (frame buffers) together as pixels,
while the hud cam is for rendering things in screen space. These are not the same thing.