Move MapToolEventBus to common lib and refactor it
MapToolEventBus is now a Real Class™ with an API. It encapsulate the Guava EventBus so that it is no longer exposed at
all (though subscribers still use @Subscribe). It is no longer faux-static, but purely instance-based. It's
singleton-ness is gained by storing it on the MapTool global class rather than obscuring this behind random
instantiations.