Remove isPersonalServer checks that simply switch between server policy and preferences
This is straightforward in the case of selecting a WalkerMetric.

For fog, it can be more subtle, but I believe this change is correct. I also think it may fix some corner cases in
ZoneView where personal servers would not respect non-GM views under the assumption that personal servers should always
use combined views.