Starting point for render instructions
Only one render instruction for now: `Noop`. This will be removed in subsequent changes and is only here for now because
the java compiler enforces that sealed interfaces have at least one implementation.

Also, in order to get the render instructions computed via `paintComponent()`, but also external calls to `renderZone()`
even when the GDX renderer is enabled, this required the addition of an `update()` method that preoduces the
instructions and which is called regardless of whether the Swing renderer is in use. `renderZone()` is now purely for
public access - and must call `update()` - while `renderZoneInternal()` is called by both `renderZone()` and
`paintComponent()` so that both the GDX renderer and regular screenshots can work.