[instruction] Text(String, Font, Rectangle2D, Color) TODO
There is also the `Label` pseudoinstruction that is analogous to `FlatImageLabel`, but not so dependent on a Swing
environment. The new `LabelFactory` is also analogous to `FlatImageLabelFactory`, but does not require a `Graphics2D` to
be provided to construct the labels.

Once all components have moved to the new `Label` record, we can remove `FlatImageLabel`.

TODO `Label` should not even be part of `RenderInstruction`, and should not have a `Decoration` enum.