Generalize AreaRenderer to handle general Shapes
Both `#fillArea(PolygonSpriteBatch, Area)` and `#drawArea(PolygonSpriteBatch, Area, boolean, float)` have been
generalized to accept `Shape` instead of `Area`. This exposed some weaknesses in the original implementation, such as
the lack of `PathIterator.SEG_CLOSE` segments when consuming the shape's `PathIterator`.

For efficiency, simple shapes such as `Ellipse2D` and `Rectangle2D` are given special treatment in `fillArea()`. This
matters particularly for small circles, which due to their tight radius make the triangulation explode into thousands of
triangles. This is by no means a complete fix - we really need to tune the triangulation to only produce as many
triangles as are needed for reasonable rendering needs.