[init-gdx] Implement a new mesher for paths TODO
This one is created whole-cloth, not taken from shapedrawer. It supports all widths, caps, and joins for `BasicStroke`,
though no support for dashes. It is also not sensitive to certain edge cases like turn backs or straight joints with no
deflection.

When the line width is large and the line length is small, there are probably subtle issues that still need to be worked
out.

TODO Need to handle single points by applying end caps without intermediate line.