[layer] Implement render layers
A render layer is a carrier for state that affects several related rendering instructions. When rendering a render
layer, all instructions are rendered onto a blank back buffer. Once all the layer's instruciton have be processed, the
back buffer is then composited with the results buffer according to the layer's configured blend mode.

Each render layer also a clipping region. But instead of sending the same clipping region redundantly for several render
layers, we instead pass all possible clipping regions along with the instruction set. The allows the renderers to use
more efficient strategies in preparing the clipping regions than could be achieved if the clips were sent along with the
render layers or even individual instructions. Render layers are preconfigured with a specific clip type that dictates
which clipping region will be applied to it. There are three clip types:
- `NoClipping`, when rendering does not require any clip to be set.
- `VisibleArea`, when rendering should be resricted to areas that the player's tokens can currently see.
- `ExposedArea`, when rendering should be restricted to areas that have fog of war revealed, i.e., that have previously
  been seen or otherwise shown to players.
- `ClearArea`, when rendering should be restricted to both the visible and exposed areas.

There is still room for significant improvements to clips. For example, instead of sending the clip areas along with
each instruction set, we could instead sync the areas to each renderer only when they change. But for now, there is not
much overhead with the clips, and until it becomes a problem it is easier to just send the latest clips.

Swing Renderer
--------------

In `ZoneRenderer`, each render layer gets its own `Graphics2D` object. The layer's clip is set directly on this object,
which avoids any cross contamination with other layers. There's not much else to clips since `Graphics2D` has direct
support for them.

TODO In `ZoneRenderer`, the above is not really respected. There is great expense in rendering to a back buffer, and we
should avoid it unless absolutely necessary. Thankfully, most layers ought to be internally blended using premultiplied
alpha, and also blended down with premultiplied alpha. Since premultiplied alpha is associated, this is equivalent to
directly actioning instructions on the results buffer. In other words, for most render layers, we can skip the back
buffer and directly render to the more efficient results buffer. What we still need to do is make this explicit rather
than defering to the various renderers to decide whether to use buffered rendering or not.

LibGDX Renderer
---------------

In the `GdxRenderer`, clips need to be converted into a form that can be used by OpenGL to implement clipping behaviour.
It does this by rendering each clip area to its own alpha-only framebuffer. Where the clip's alpha is 0, the result's
pixel will be masked and not visible. But where the alpha is 1, the result's pixel will be rendered to screen. In
principle, we can have intermediate alpha values as well. But this is unlikely to be used directly, and is more likely
to arise in cases where we are doing some kind of antialiasing or other processing that might blend neighbouring pixel
values together.

The `GdxRenderer` also has solid support for render layers. We rely on the previously introduced system of three
framebuffers (results, back, and spare buffers), but now layers are handled with much more consistency. This makes state
management much easier, since we know each layer is always rendering to `backBuffer` which will then be blended with
`resultsBuffer` into `spareBuffer`, and then `resultsBuffer` and `spareBuffer` will be swapped. Within render layers,
there are no more assumptions about which methods have to restore which buffers as the current framebuffer, or otherwise
make assumptions about where they are rendering to.

The existing `environmentalLightingShader.{vsh,fsh}` has been expanded into `layerShader.{vsh,fsh}`, which makes several
key improvements:
1. An additional uniform texture input to act as the clip. This ultimately is bound to the framebuffer associated with
   the render layer's clip type.
2. An additional uniform enumerator that allows selecting the blend mode. For now, only premultiplied alpha blending is
   supported, but additional blend modes will be added later for straight alpha and environmental lighting.
3. Does not use the texture lookup coordinate `v_texCoords` to index the screen-sized mask texture.

To make the shader easier to use, it has been encapsulated into `LayerShader`.

CONTINUE Switch to more dynamic layering

The instructions are now a flat list rather than being divided by render layer. Render layers can be started and stopped
at anytime, with the structure of `InstructionSetBuilder` guaranteeing proper balance.

`RenderLayerBuilder` is no more. Since instructions are flat, that functionality has been moved into
`InstructionSetBuilder`.

`ZoneCompositor` maintains its own list of instructions now. It currently still copies them at the end of
`processInstructions()`, but smarter management will now be possible in the future.

CONTINUE Add layer timers with blit timers, and corresponding increments