[grid] Reimplement grid rendering via compositor TODO
The compositor produces a `Path2D` representing the grid pattern that needs to be drawn on screen, then passes that to
the renderers as a sequence of `Stroke` instructions. The Swing and LibGDX renderers no longer need any logic dedicated
to rendering grids beyond their generic `Stroke` handling.

This naturally fixes a bug in LibGDX's `GridRenderer` where the pooled `Matrix4` is not reset before being used, but is
instead reset before being freed. If any other component would have used a pooled `Matrix4`, this would have broken the
`GridRenderer.`

TODO Need to incorporate the new GridRenderer stuff while not eliminating the old.