[lights] Reimplement lights and auras using compositor
For the LibGDX renderer, Lights and auras don't explicitly manage the back buffer, but instead rely on our generalized
blending shader. This comes with two new blend modes:
1. For overtop lighting, straight alpha compositing is needed instead of premultiplied since the lights are first
   blending using the screen mode which does not produce premultiplied results.
2. For environmental lights, the brighten blend mode is needed.

Because of the generalized clipping support, lights and auras are naturally clipped to the visible region.

There will be no more need for the `LightsRenderer` for auras or lights since they are represented by standard `Fill`
instructions now.