Update initiative list to not horizontally scroll
The cell renderer now sets its width according to the JList's width so it never overflows. The JList itself is
explicitly set to no longer scroll.

The layout of the cell has been replaced with a GridLayoutManager-based solution, instead of MigLayout. This gives us
better control over how each element sizes itself compared to others. The end result is effectively the same as the
original, but with the ability for labels to respect the width and ellipsify on overflow

TODO When the second line is not enabled, don't show that label