Token copy constructor does not keep exposedAreaGUID
The ID is only kept if the overload is called with `keepIds == true`.

In places that copied tokens and checked if hte exposedAreaGUID needed to be changed: most of these places either know
there will be a collision and should always have it changed, or know there will be no collision. So there is no more
need to gather all existing IDs and then check after the fact.