Clean up Zone copying
In `#readResolve()`, make sure the `#initiativeList` is non-null, so that the copy constructor does not have to detect
this case and modify the copied `Zone`.

For the copy constructor itself, focus on copying the data directly rather than using higher-level operations. This
avoids pumping events for various operations when copying the zones, and also makes everything more direct.